\name{pbdPAPI-package}
\alias{pbdPAPI-package}
\docType{package}
\title{PAPI}
\description{
  PAPI
}
\details{
\tabular{ll}{
  Package: \tab pbdPAPI\cr
  Type: \tab Package\cr
  License: \tab BSD 2-Clause\cr
  LazyLoad: \tab yes\cr
}
}
\references{
  Programming with Big Data in R Website:
  \url{http://r-pbd.org/}
  PAPI website: \url{http://icl.cs.utk.edu/PAPI/}
}
\author{
  Drew Schmidt \email{schmidt AT math.utk.edu}, Christian
  Hckendorf, Wei-Chen Chen, George Ostrouchov, and Pragneshkumar 
  Patel.
}
\keyword{Package}
