\name{pbdPAPI Global Environment}
\alias{.__pbdPAPIEnv}
\docType{data}
\title{Global Environment for the pbdPAPI Package}
\description{
  The environment for the pbdBASE package where "global" variables are
  stored.
}
\details{
  The \code{.__pbdPAPI_avail} object is stored in the \code{.__pbdPAPIEnv}
  environment.
}
