% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cachebench.r
\name{cachebench}
\alias{cachebench}
\title{cachebench}
\usage{
cachebench(..., nreps = 10)
}
\arguments{
\item{...}{Expressions to be benchmarked.}

\item{nreps}{Number of replications for each expression.}
}
\value{

}
\description{
A benchmarker for
}
\details{
Currently only total cache misses are profiled.  Future versions
will allow the full generality of the \code{systme.cache()}
function.
}
\examples{
\dontrun{
cachebench(rnorm(1e4), rnorm(2e4), rnorm(3e4))
}

}

