% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flopsbench.r
\name{flopsbench}
\alias{flopsbench}
\title{flopsbench}
\usage{
flopsbench(..., data)
}
\arguments{
\item{...}{Expressions to be benchmarked.}

\item{data}{A list of data for the provided functions.}
}
\value{
A dataframe (class "flopsbench") containng the timings and some
basic summary statistics.
}
\description{
A benchmarker for measuring flops (floating point operations per
second) of one or more functions.
}
\examples{
\dontrun{
data <- list(runif(1e4), runif(2e4), runif(3e4))
flopsbench(sqrt, exp, data=data)
}

}

