% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipcm.r
\name{ipcm.cache}
\alias{ipcm.cache}
\title{IPCM Cache}
\usage{
ipcm.cache(expr, type = "all", level = NULL, gcFirst = TRUE,
  burnin = TRUE)
}
\arguments{
\item{expr}{A valid R expression to be profiled.}

\item{type}{The type of cache event to profile; options are "miss", "hit",
and "all".}

\item{level}{The cache level to profile; options are "2" and "3".}

\item{gcFirst}{logical; determines if garbage collection should be called
before profiling.}

\item{burnin}{logical; determines if the function should first be evaluated
with an empty expression.}
}
\value{
The results of the requested PAPI events are returned, in a named
list, with values stored in double precision.
}
\description{
Measuring cache misses and hits
}
\details{
This function measures the evaluation of the provided expression, expr, for
different classes of cache events additionally restricted to misses or hits.
The additional arguments, gcFirst and burnin, can be toggled to better
simulate the target context.
}
\examples{
\dontrun{
library(pbdPAPI)

ipcm.cache(1+1, type="miss", level="2")
}

}
\keyword{programming}

