% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/envir.r
\docType{data}
\name{pbdPAPI Global Environment}
\alias{.__pbdPAPIEnv}
\alias{pbdPAPI Global Environment}
\title{Global Environment for the pbdPAPI Package}
\format{\preformatted{<environment: 0x27f8340> 
}}
\usage{
.__pbdPAPIEnv
}
\description{
The environment for the pbdBASE package where package state is stored.
}
\details{
The \code{.__pbdPAPI_avail} object is stored in the \code{.__pbdPAPIEnv}
environment.
}
\keyword{datasets}

