% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbdPAPI-package.r
\docType{package}
\name{pbdPAPI-package}
\alias{pbdPAPI-package}
\title{Performance Analysis Tools for R}
\description{
pbdPAPI is an R package which offers a set of R bindings for the Performance
Application Programming Interface (PAPI).  PAPI is a high-level C library
for gathering information from hardware performance counters.  PAPI measures
events like cache misses, flops, etc.
}
\details{
\tabular{ll}{ 
   Package: \tab pbdPAPI\cr 
   Type: \tab Package\cr License: \tab BSD 2-Clause\cr 
}
}
\author{
Drew Schmidt \email{schmidt AT math.utk.edu}, Christian Hckendorf,
and Wei-Chen Chen.
}
\references{
Programming with Big Data in R Website: \url{http://r-pbd.org/}
PAPI website: \url{http://icl.cs.utk.edu/PAPI/}
}

