% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.papi_cache}
\alias{plot.papi_cache}
\title{Plots PAPI objects}
\usage{
\method{plot}{papi_cache}(x, ..., title, opnames, color = FALSE,
  show.opnames = TRUE, facet.by = "opnames")
}
\arguments{
\item{x}{PAPI object.}

\item{...}{Additional objects.}

\item{title}{Optional argument for adding a title to the plot.}

\item{opnames}{An optional argument to specify different names for the
expressions/operations used to generate the profiler data.}

\item{color}{Logical; indicates whether groups should be colored differently
in the case of plotting multiple objects}

\item{show.opnames}{Logical; indicates whether or not expression/operation names should be
shown.}

\item{facet.by}{Choice to facet cache plots by the different expressions/operations
(\code{facet.by="opnames"}), or by the cache level (\code{facet.by="level"}).}
}
\description{
Plots PAPI objects
}

