% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cachebench.r
\name{print-cachebench}
\alias{print-cachebench}
\alias{print.cachebench}
\title{Prints cachebench objects}
\usage{
\method{print}{cachebench}(x, ...)
}
\arguments{
\item{x}{A papi_output object to print.}

\item{...}{Extra arguments (ignored).}
}
\description{
Prints cachebench objects
}

