% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.PAPI}
\alias{print.IPCM}
\alias{print.PAPI}
\alias{print.papi_output}
\title{Prints PAPI objects}
\usage{
\method{print}{PAPI}(x, ...)

\method{print}{IPCM}(x, ...)

\method{print}{papi_output}(x, digits = 5)
}
\arguments{
\item{x}{PAPI object.}

\item{...}{ignored}

\item{digits}{Number of digits to print.}
}
\description{
Prints PAPI objects
}

