% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.r
\name{[.papi_output}
\alias{[.papi_output}
\title{Subsets PAPI outputs.}
\usage{
\method{[}{papi_output}(x, i)
}
\arguments{
\item{x}{PAPI object.}

\item{i}{Index.

rdname subset}
}
\description{
Subsets PAPI outputs.
}

