% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.r
\name{system.cache}
\alias{system.cache}
\title{Cache}
\usage{
system.cache(expr, type = "miss", events = "total", gcFirst = TRUE,
  burnin = TRUE)
}
\arguments{
\item{expr}{A valid R expression to be profiled.}

\item{type}{The type of cache event to profile; options are "miss", "hit",
"access", and "read".}

\item{events}{The class of cache events to profile; options are "total",
"data", "instruction", "all", "l1.all", "l2.all", "l3.all", "l1.ratio",
"l2.ratio", and "l3.ratio".}

\item{gcFirst}{logical; determines if garbage collection should be called
before profiling.}

\item{burnin}{logical; determines if the function should first be evaluated
with an empty expression.}
}
\value{
The results of the requested PAPI events are returned, in a named
list, with values stored in double precision.
}
\description{
Measuring cache misses, hits, accesses, and reads.
}
\details{
This function measures the evaluation of the provided expression, expr, for
different classes of cache events additionally restricted to misses, hits,
accesses, or reads. The additional arguments, gcFirst and burnin, can be
toggled to better simulate the target context.
}
\examples{
\dontrun{
library(pbdPAPI)

system.cache(1+1, type="miss", events="l2.all")
}

}
\keyword{programming}

