% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flips.r
\name{system.flips}
\alias{system.flips}
\title{flips}
\usage{
system.flips(expr, gcFirst = TRUE, burnin = TRUE)
}
\arguments{
\item{expr}{A valid R expression to be profiled.}

\item{gcFirst}{logical; determines if garbage collection should be called
before profiling.}

\item{burnin}{logical; determines if the function should first be evaluated
with an empty expression.}
}
\value{
The return is a list consisting of: 
\tabular{ll}{ 
   \code{real_time} \tab real time spent evaluating expression \cr 
   \code{proc_time} \tab total process time spent evaluating expression \cr 
   \code{flpins} \tab FLoating Point INStructions (count) \cr 
   \code{mflips} \tab Mflips (flpins per second) \cr 
}
}
\description{
Measuring floating point instructions.
}
\details{
This function measures the evaluation of the provided expression, expr, for
real time, process time, number of floating point instructions, and floating
point instruction rate. The additional arguments, gcFirst and burnin, can be
toggled to better simulate the target context.
}
\examples{
\dontrun{
library(pbdPAPI)

system.flips(1+1)
}

}
\keyword{programming}

