% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idle.r
\name{system.idle}
\alias{system.idle}
\title{Idle}
\usage{
system.idle(expr, events = "float", gcFirst = TRUE, burnin = TRUE)
}
\arguments{
\item{expr}{A valid R expression to be profiled.}

\item{events}{The class of processing unit to profile; options are "float",
"int", "numeric", "branch", "load", and "all".}

\item{gcFirst}{logical; determines if garbage collection should be called
before profiling.}

\item{burnin}{logical; determines if the function should first be evaluated
with an empty expression.}
}
\value{
The results of the requested PAPI events are returned, in a named
list, with values stored in double precision.
}
\description{
Measuring idle cycles.
}
\details{
This function measures the evaluation of the provided expression, expr, for
the number of cycles that specific specialized processing units are idle.
The additional arguments, gcFirst and burnin, can be toggled to better
simulate the target context.
}
\examples{
\dontrun{
library(pbdPAPI)

system.idle(1+1, events="float")
}

}
\keyword{programming}

