% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cachebench.r
\name{cachebench}
\alias{cachebench}
\title{cachebench}
\usage{
cachebench(..., type = "miss", events = "total", nreps = 10)
}
\arguments{
\item{...}{Expressions to be benchmarked.}

\item{type, events}{See \code{?system.cache()}.}

\item{nreps}{Number of replications for each expression.}
}
\value{
A dataframe (class "cachebench") containng the timings and some
basic summary statistics.
}
\description{
A benchmarker for measuring cache misses of multiple runs of
one or more functions.
}
\details{
Currently only total cache misses are profiled.  Future versions
will allow the full generality of the \code{systme.cache()}
function.
}
\examples{
\dontrun{
pbdPAPI::cachebench(rnorm(1e4), rnorm(2e4), rnorm(3e4))
}

}

