% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envir.r
\docType{data}
\name{pbdPAPI Global Environment}
\alias{.__pbdPAPIEnv}
\alias{pbdPAPI Global Environment}
\title{Global Environment for the pbdPAPI Package}
\format{An object of class \code{environment} of length 0.}
\usage{
.__pbdPAPIEnv
}
\description{
The environment for the pbdBASE package where package state is stored.
}
\details{
The \code{.__pbdPAPI_avail} object is stored in the \code{.__pbdPAPIEnv}
environment.
}
\keyword{datasets}

