% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.r
\name{print.papi_output}
\alias{print.papi_output}
\title{Prints PAPI objects}
\usage{
\method{print}{papi_output}(x, ..., digits = 5)
}
\arguments{
\item{x}{A papi_output object to print.}

\item{...}{Extra arguments (ignored).}

\item{digits}{Number of digits to print.}
}
\description{
Prints PAPI objects
}

