% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncounters.r
\name{system.ncounters}
\alias{system.ncounters}
\title{Num.Counters}
\usage{
system.ncounters()
}
\value{
The number of counters available on the system.
}
\description{
Number of counters available on the system.
}
\details{
This function returns the number of counters available on the
system.  A return value of zero indicates that your system
does not support profiling by hardware event counters.
}
\examples{
\dontrun{
library(pbdPAPI)

system.ncounters()
}

}
\keyword{programming}

